UX-565 Using `css` prop instead of `styled.` helper function
